Changed benchmark file structure.
Benchmarks are now classified as unary or binary, corresponding to the
type of operation being benchmarked. These are contained in
`benchmark_unary` and `benchmark_binary`. Operations to be benchmarked
are automatically discovered.
